requirements.txt
# Face Verification System - Requirements

# Web Framework & API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ML & Computer Vision
torch==2.1.1
torchvision==0.16.1
tensorflow==2.14.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
scikit-image==0.22.0
scikit-learn==1.3.2

# Face Recognition & Detection
insightface==0.7.3
retinaface==0.0.17
dlib==19.24.2
mediapipe==0.10.9

# Image Processing & Quality
Pillow==10.1.0
imageio==2.33.1
brisque==0.0.11

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
cryptography==41.0.7
alembic==1.12.1

# Caching & Message Queue
redis==5.0.1
kafka-python==2.0.2

# Security & Encryption
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.2.0

# AWS/Cloud Integration
boto3==1.29.7
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Data Processing
pandas==2.1.3
pyarrow==14.0.0

# Configuration Management
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Code Quality (dev)
black==23.12.0
flake8==6.1.0
pylint==3.0.3
isort==5.13.2
mypy==1.7.1

# Documentation (dev)
mkdocs==1.5.3
mkdocs-material==9.5.3

# Utilities
tqdm==4.66.1
requests==2.31.0
urllib3==2.1.0

# Performance Testing (dev)
locust==2.17.0
psutil==5.9.6
